requirements_linux.txt
torchmetrics==0.9.3
torchsummary==1.5.1
absl-py==1.2.0
accelerate==0.12.0
addict==2.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.8.0
bleach==5.0.1
cachetools==5.2.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
fire==0.4.0
flake8==5.0.4
fonttools==4.37.1
fvcore==0.1.5.post20220512
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
idna==3.3
imageio==2.21.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
iopath==0.1.10
ipykernel==6.15.2
ipython
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-client==7.3.5
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
llvmlite==0.31.0
lxml==4.9.1
lyft-dataset-sdk==0.0.8
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mmcls==0.23.2
mmcv-full==1.3.14
mmdet==2.14.0
mmsegmentation==0.14.1
mypy-extensions==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.2
notebook==6.4.12
numba==0.48.0
numpy==1.19.0
nuscenes-devkit==1.1.9
oauthlib==3.2.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.2
plotly==5.10.0
pluggy==1.0.0
plyfile==0.7.4
portalocker==2.5.1
prettytable==3.4.1
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.13.0
pyparsing==3.0.9
pyquaternion==0.9.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.18.3
scikit-learn
scipy==1.9.1
Send2Trash==1.8.0
Shapely==1.8.4
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
tabulate==0.8.10
tenacity==8.0.1
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.15.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tifffile
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.3.0
trimesh==2.35.39
typing-extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yacs==0.1.8
yapf==0.32.0
zipp==3.8.1
torch_tb_profiler